@@ -1,43 +1,39 @@
-// swift-tools-version:5.6
+// swift-tools-version:6.0
 import PackageDescription
-
-// This package contains a vendored copy of BoringSSL. For ease of tracking
-// down problems with the copy of BoringSSL in use, we include a copy of the
-// commit hash of the revision of BoringSSL included in the given release.
-// This is also reproduced in a file called hash.txt in the
-// Sources/CCryptoBoringSSL directory. The source repository is at
-// https://boringssl.googlesource.com/boringssl.
-//
-// BoringSSL Commit: 58a318edc892a595a5b043359a5d441869158699
 
 let package = Package(
     name: "jwt-kit",
     platforms: [
-        .macOS(.v10_15),
-        .iOS(.v13),
-        .tvOS(.v13),
-        .watchOS(.v6),
+        .macOS(.v13),
+        .iOS(.v15),
+        .tvOS(.v15),
+        .watchOS(.v8),
     ],
     products: [
-        .library(name: "JWTKit", targets: ["JWTKit"]),
-        /* This target is used only for symbol mangling. It's added and removed automatically because it emits build warnings. MANGLE_START
-        .library(name: "CJWTKitBoringSSL", type: .static, targets: ["CJWTKitBoringSSL"]),
-        MANGLE_END */
+        .library(name: "JWTKit", targets: ["JWTKit"])
     ],
     dependencies: [
-        .package(url: "https://github.com/apple/swift-crypto.git", "2.0.0" ..< "4.0.0"),
+        .package(url: "https://github.com/apple/swift-crypto.git", "3.8.0"..<"5.0.0"),
+        .package(url: "https://github.com/apple/swift-certificates.git", from: "1.2.0"),
+        .package(url: "https://github.com/apple/swift-log.git", from: "1.0.0"),
     ],
     targets: [
-        .target(name: "CJWTKitBoringSSL"),
-        .target(name: "JWTKit", dependencies: [
-            .target(name: "CJWTKitBoringSSL"),
-            .product(name: "Crypto", package: "swift-crypto"),
-        ]),
-        .testTarget(name: "JWTKitTests", dependencies: [
-            .target(name: "JWTKit"),
-        ]),
-    ],
-    cxxLanguageStandard: .cxx11
+        .target(
+            name: "JWTKit",
+            dependencies: [
+                .product(name: "Crypto", package: "swift-crypto"),
+                .product(name: "_CryptoExtras", package: "swift-crypto"),
+                .product(name: "X509", package: "swift-certificates"),
+                .product(name: "Logging", package: "swift-log"),
+            ]
+        ),
+        .testTarget(
+            name: "JWTKitTests",
+            dependencies: [
+                "JWTKit"
+            ]
+        ),
+    ]
 )
 
 import Foundation
