@@ -1,7 +1,7 @@
 /// A claim is a codable, top-level property of a JWT payload. Multiple claims form a payload.
 /// Some claims, such as expiration claims, are inherently verifiable. Each claim able to verify
 /// itself provides an appropriate method for doing so, depending on the specific claim.
-public protocol JWTClaim: Codable {
+public protocol JWTClaim: Codable, Sendable {
     /// The associated value type.
     associatedtype Value: Codable
 
