@@ -1,34 +1,47 @@
-import Foundation
 import Crypto
 
-internal struct EdDSASigner: JWTAlgorithm {
-    let key: EdDSAKey
+#if !canImport(Darwin)
+    import FoundationEssentials
+#else
+    import Foundation
+#endif
+
+struct EdDSASigner: JWTAlgorithm, Sendable {
+    let publicKey: EdDSA.PublicKey
+    let privateKey: EdDSA.PrivateKey?
     let name = "EdDSA"
-    
-    func sign<Plaintext>(_ plaintext: Plaintext) throws -> [UInt8] where Plaintext : DataProtocol {                
-        guard let privateKey = key.privateKey else {
-            throw  JWTError.signingAlgorithmFailure(EdDSAError.privateKeyMissing)
-        }
-        
-        switch key.curve {
-        case .ed25519:
-            return try Curve25519.Signing.PrivateKey(
-                rawRepresentation: privateKey
-            ).signature(
-                for: plaintext
-            ).copyBytes()
+
+    init(key: some EdDSAKey) {
+        switch key {
+        case let key as EdDSA.PrivateKey:
+            self.privateKey = key
+            self.publicKey = key.publicKey
+        case let key as EdDSA.PublicKey:
+            self.publicKey = key
+            self.privateKey = nil
+        default:
+            // This should never happen
+            fatalError("Unexpected key type: \(type(of: key))")
         }
     }
-    
-    func verify<Signature, Plaintext>(_ signature: Signature, signs plaintext: Plaintext) throws -> Bool where Signature : DataProtocol, Plaintext : DataProtocol {
-        switch key.curve {
+
+    func sign(_ plaintext: some DataProtocol) throws -> [UInt8] {
+        guard let privateKey else {
+            throw JWTError.signingAlgorithmFailure(EdDSAError.privateKeyMissing)
+        }
+
+        switch privateKey.curve.backing {
         case .ed25519:
-            return try Curve25519.Signing.PublicKey(
-                rawRepresentation: key.publicKey
-            ).isValidSignature(
-                signature,
-                for: plaintext
-            )
+            return try Curve25519.Signing.PrivateKey(rawRepresentation: privateKey.rawRepresentation)
+                .signature(for: plaintext).copyBytes()
+        }
+    }
+
+    func verify(_ signature: some DataProtocol, signs plaintext: some DataProtocol) throws -> Bool {
+        switch publicKey.curve.backing {
+        case .ed25519:
+            try Curve25519.Signing.PublicKey(rawRepresentation: publicKey.rawRepresentation)
+                .isValidSignature(signature, for: plaintext)
         }
     }
 }
