@@ -2,10 +2,10 @@
 ///
 /// A JSON object that represents a set of JWKs.
 /// Read specification (RFC 7517) https://tools.ietf.org/html/rfc7517.
-public struct JWKS: Codable {
+public struct JWKS: Codable, Sendable {
     /// All JSON Web Keys
     public var keys: [JWK]
-    
+
     public init(keys: [JWK]) {
         self.keys = keys
     }
