@@ -1,44 +1,192 @@
-import Foundation
+#if !canImport(Darwin)
+    import FoundationEssentials
+#else
+    import Foundation
+#endif
 
-public enum JWTError: Error, CustomStringConvertible, LocalizedError {
-    case claimVerificationFailure(name: String, reason: String)
-    case signingAlgorithmFailure(Error)
-    case malformedToken
-    case signatureVerifictionFailed
-    case missingKIDHeader
-    case unknownKID(JWKIdentifier)
-    case invalidJWK
-    case invalidBool(String)
-    case generic(identifier: String, reason: String)
+/// JWT error type.
+public struct JWTError: Error, Sendable, Equatable {
+    public struct ErrorType: Sendable, Hashable, CustomStringConvertible, Equatable {
+        enum Base: String, Sendable, Equatable {
+            case claimVerificationFailure
+            case signingAlgorithmFailure
+            case malformedToken
+            case signatureVerificationFailed
+            case missingKIDHeader
+            case unknownKID
+            case invalidJWK
+            case invalidBool
+            case noKeyProvided
+            case missingX5CHeader
+            case invalidX5CChain
+            case invalidHeaderField
+            case unsupportedCurve
+            case generic
+        }
 
-    public var reason: String {
-        switch self {
-        case .claimVerificationFailure(let name, let reason):
-            return "\(name) claim verification failed: \(reason)"
-        case .signingAlgorithmFailure(let error):
-            return "signing algorithm error: \(error)"
-        case .malformedToken:
-            return "malformed JWT"
-        case .signatureVerifictionFailed:
-            return "signature verification failed"
-        case .missingKIDHeader:
-            return "missing kid field in header"
-        case .unknownKID(let kid):
-            return "unknown kid: \(kid)"
-        case .invalidJWK:
-            return "invalid JWK"
-        case .invalidBool(let str):
-            return "invalid boolean value: \(str)"
-        case .generic(let identifier, let reason):
-                return "missing '\(identifier). \(reason)"
+        let base: Base
+
+        private init(_ base: Base) {
+            self.base = base
+        }
+
+        public static let claimVerificationFailure = Self(.claimVerificationFailure)
+        public static let signingAlgorithmFailure = Self(.signingAlgorithmFailure)
+        public static let signatureVerificationFailed = Self(.signatureVerificationFailed)
+        public static let missingKIDHeader = Self(.missingKIDHeader)
+        public static let malformedToken = Self(.malformedToken)
+        public static let unknownKID = Self(.unknownKID)
+        public static let invalidJWK = Self(.invalidJWK)
+        public static let invalidBool = Self(.invalidBool)
+        public static let noKeyProvided = Self(.noKeyProvided)
+        public static let missingX5CHeader = Self(.missingX5CHeader)
+        public static let invalidX5CChain = Self(.invalidX5CChain)
+        public static let invalidHeaderField = Self(.invalidHeaderField)
+        public static let unsupportedCurve = Self(.unsupportedCurve)
+        public static let generic = Self(.generic)
+
+        public var description: String {
+            base.rawValue
         }
     }
 
-    public var description: String {
-        return "JWTKit error: \(self.reason)"
+    private struct Backing: Sendable, Equatable {
+        fileprivate let errorType: ErrorType
+        fileprivate let name: String?
+        fileprivate let reason: String?
+        fileprivate let underlying: Error?
+        fileprivate let kid: JWKIdentifier?
+        fileprivate let identifier: String?
+        fileprivate let failedClaim: (any JWTClaim)?
+        fileprivate var curve: (any ECDSACurveType)?
+
+        init(
+            errorType: ErrorType,
+            name: String? = nil,
+            reason: String? = nil,
+            underlying: Error? = nil,
+            kid: JWKIdentifier? = nil,
+            identifier: String? = nil,
+            failedClaim: (any JWTClaim)? = nil,
+            curve: (any ECDSACurveType)? = nil
+        ) {
+            self.errorType = errorType
+            self.name = name
+            self.reason = reason
+            self.underlying = underlying
+            self.kid = kid
+            self.identifier = identifier
+            self.failedClaim = failedClaim
+            self.curve = curve
+        }
+
+        static func == (lhs: JWTError.Backing, rhs: JWTError.Backing) -> Bool {
+            lhs.errorType == rhs.errorType
+        }
     }
-    
-    public var errorDescription: String? {
-        return self.description
+
+    private var backing: Backing
+
+    public var errorType: ErrorType { backing.errorType }
+    public var name: String? { backing.name }
+    public var reason: String? { backing.reason }
+    public var underlying: (any Error)? { backing.underlying }
+    public var kid: JWKIdentifier? { backing.kid }
+    public var identifier: String? { backing.identifier }
+    public var failedClaim: (any JWTClaim)? { backing.failedClaim }
+    public var curve: (any ECDSACurveType)? { backing.curve }
+
+    private init(backing: Backing) {
+        self.backing = backing
+    }
+
+    private init(errorType: ErrorType) {
+        self.backing = .init(errorType: errorType)
+    }
+
+    public static func claimVerificationFailure(failedClaim: (any JWTClaim)?, reason: String) -> Self {
+        .init(backing: .init(errorType: .claimVerificationFailure, reason: reason, failedClaim: failedClaim))
+    }
+
+    public static func signingAlgorithmFailure(_ error: Error) -> Self {
+        .init(backing: .init(errorType: .signingAlgorithmFailure, underlying: error))
+    }
+
+    public static func malformedToken(reason: String) -> Self {
+        .init(backing: .init(errorType: .malformedToken, reason: reason))
+    }
+
+    public static let signatureVerificationFailed = Self(errorType: .signatureVerificationFailed)
+
+    public static let missingKIDHeader = Self(errorType: .missingKIDHeader)
+
+    public static func unknownKID(_ kid: JWKIdentifier) -> Self {
+        .init(backing: .init(errorType: .unknownKID, kid: kid))
+    }
+
+    public static func invalidJWK(reason: String) -> Self {
+        .init(backing: .init(errorType: .invalidJWK, reason: reason))
+    }
+
+    public static func invalidBool(_ name: String) -> Self {
+        .init(backing: .init(errorType: .invalidBool, name: name))
+    }
+
+    public static let noKeyProvided = Self(errorType: .noKeyProvided)
+
+    public static let missingX5CHeader = Self(errorType: .missingX5CHeader)
+
+    public static func invalidX5CChain(reason: String) -> Self {
+        .init(backing: .init(errorType: .invalidX5CChain, reason: reason))
+    }
+
+    public static func invalidHeaderField(reason: String) -> Self {
+        .init(backing: .init(errorType: .invalidHeaderField, reason: reason))
+    }
+
+    public static func unsupportedCurve(curve: any ECDSACurveType) -> Self {
+        .init(backing: .init(errorType: .unsupportedCurve, curve: curve))
+    }
+
+    public static func generic(identifier: String, reason: String) -> Self {
+        .init(backing: .init(errorType: .generic, reason: reason))
+    }
+
+    public static func == (lhs: JWTError, rhs: JWTError) -> Bool {
+        lhs.backing == rhs.backing
+    }
+}
+
+extension JWTError: CustomStringConvertible {
+    public var description: String {
+        var result = #"JWTKitError(errorType: \#(self.errorType)"#
+
+        if let name {
+            result.append(", name: \(String(reflecting: name))")
+        }
+
+        if let failedClaim {
+            result.append(", failedClaim: \(String(reflecting: failedClaim))")
+        }
+
+        if let reason {
+            result.append(", reason: \(String(reflecting: reason))")
+        }
+
+        if let underlying {
+            result.append(", underlying: \(String(reflecting: underlying))")
+        }
+
+        if let kid {
+            result.append(", kid: \(String(reflecting: kid))")
+        }
+
+        if let identifier {
+            result.append(", identifier: \(String(reflecting: identifier))")
+        }
+
+        result.append(")")
+
+        return result
     }
 }
