@@ -1,15 +1,18 @@
-import Foundation
+#if !canImport(Darwin)
+    import FoundationEssentials
+#else
+    import Foundation
+#endif
 
-public protocol JWTJSONDecoder {
+public protocol JWTJSONDecoder: Sendable {
     func decode<T: Decodable>(_: T.Type, from string: Data) throws -> T
 }
 
-public protocol JWTJSONEncoder {
+public protocol JWTJSONEncoder: Sendable {
     func encode<T: Encodable>(_ value: T) throws -> Data
 }
 
 extension JSONDecoder: JWTJSONDecoder {}
-
 extension JSONEncoder: JWTJSONEncoder {}
 
 extension JSONDecoder.DateDecodingStrategy {
@@ -31,7 +34,7 @@ extension JSONEncoder.DateEncodingStrategy {
 }
 
 extension JWTJSONEncoder where Self == JSONEncoder {
-    static var defaultForJWT: any JWTJSONEncoder {
+    public static var defaultForJWT: any JWTJSONEncoder {
         let encoder = JSONEncoder()
         encoder.dateEncodingStrategy = .secondsSince1970
         return encoder
@@ -39,7 +42,7 @@ extension JWTJSONEncoder where Self == JSONEncoder {
 }
 
 extension JWTJSONDecoder where Self == JSONDecoder {
-    static var defaultForJWT: any JWTJSONDecoder {
+    public static var defaultForJWT: any JWTJSONDecoder {
         let decoder = JSONDecoder()
         decoder.dateDecodingStrategy = .secondsSince1970
         return decoder
