@@ -1,14 +1,18 @@
-import Foundation
+#if !canImport(Darwin)
+    import FoundationEssentials
+#else
+    import Foundation
+#endif
 
-public protocol JWTUnixEpochClaim: JWTClaim where Value == Date { }
+public protocol JWTUnixEpochClaim: JWTClaim where Value == Date {}
 
 extension JWTUnixEpochClaim {
     /// See `Decodable`.
     public init(from decoder: Decoder) throws {
         let container = try decoder.singleValueContainer()
-        self.init(value: try container.decode(Date.self))
+        try self.init(value: container.decode(Date.self))
     }
-    
+
     /// See `Encodable`.
     public func encode(to encoder: Encoder) throws {
         var container = encoder.singleValueContainer()
