@@ -1,17 +1,21 @@
-import Foundation
+#if !canImport(Darwin)
+    import FoundationEssentials
+#else
+    import Foundation
+#endif
 
 /// The "nbf" (not before) claim identifies the time before which the JWT
-/// MUST NOT be accepted for processing.  The processing of the "nbf"
+/// MUST NOT be accepted for processing. The processing of the "nbf"
 /// claim requires that the current date/time MUST be after or equal to
-/// the not-before date/time listed in the "nbf" claim.  Implementers MAY
+/// the not-before date/time listed in the "nbf" claim. Implementers MAY
 /// provide for some small leeway, usually no more than a few minutes, to
-/// account for clock skew.  Its value MUST be a number containing a
-/// NumericDate value.  Use of this claim is OPTIONAL.
+/// account for clock skew. Its value MUST be a number containing a
+/// NumericDate value. Use of this claim is OPTIONAL.
 public struct NotBeforeClaim: JWTUnixEpochClaim, Equatable {
-    /// See `JWTClaim`.
+    /// See ``JWTClaim``.
     public var value: Date
 
-    /// See `JWTClaim`.
+    /// See ``JWTClaim``.
     public init(value: Date) {
         self.value = value
     }
@@ -20,7 +24,7 @@ public struct NotBeforeClaim: JWTUnixEpochClaim, Equatable {
     public func verifyNotBefore(currentDate: Date = .init()) throws {
         switch value.compare(currentDate) {
         case .orderedDescending:
-            throw JWTError.claimVerificationFailure(name: "nbf", reason: "too soon")
+            throw JWTError.claimVerificationFailure(failedClaim: self, reason: "too soon")
         case .orderedAscending, .orderedSame:
             break
         }
