@@ -1,5 +1,7 @@
-internal enum ECDSAError: Error {
+enum ECDSAError: Error {
     case newKeyByCurveFailure
     case generateKeyFailure
     case signFailure
+    case noPublicKey
+    case noPrivateKey
 }
