@@ -1,4 +1,4 @@
-internal enum EdDSAError: Error {
+enum EdDSAError: Error {
     case publicAndPrivateKeyMissing
     case privateKeyMissing
     case publicKeyMissing
