@@ -1,25 +1,43 @@
-import Foundation
+#if !canImport(Darwin)
+    import FoundationEssentials
+#else
+    import Foundation
+#endif
 
-extension DataProtocol {
-    func base64URLDecodedBytes() -> [UInt8] {
-        return Data(base64Encoded: Data(self.copyBytes()).base64URLUnescaped())?.copyBytes() ?? []
-    }
+extension String {
+    package func base64URLDecodedData() -> Data? {
+        var base64URL = replacingOccurrences(of: "-", with: "+")
+            .replacingOccurrences(of: "_", with: "/")
 
-    func base64URLEncodedBytes() -> [UInt8] {
-        return Data(self.copyBytes()).base64EncodedData().base64URLEscaped().copyBytes()
+        base64URL.append(contentsOf: "===".prefix((4 - (base64URL.count & 3)) & 3))
+
+        return Data(base64Encoded: base64URL)
     }
 }
 
-/// MARK: Data Escape
-private extension Data {
+extension DataProtocol {
+    package func base64URLDecodedBytes() -> [UInt8] {
+        Data(base64Encoded: Data(copyBytes()).base64URLUnescaped())?.copyBytes() ?? []
+    }
+
+    package func base64URLEncodedBytes() -> [UInt8] {
+        Data(copyBytes()).base64EncodedData().base64URLEscaped().copyBytes()
+    }
+}
+
+// MARK: Data Escape
+
+extension Data {
     /// Converts base64-url encoded data to a base64 encoded data.
     ///
     /// https://tools.ietf.org/html/rfc4648#page-7
-    mutating func base64URLUnescape() {
-        for (i, byte) in self.enumerated() {
-            switch byte {
-            case 0x2D: self[self.index(self.startIndex, offsetBy: i)] = 0x2B
-            case 0x5F: self[self.index(self.startIndex, offsetBy: i)] = 0x2F
+    fileprivate mutating func base64URLUnescape() {
+        for idx in self.indices {
+            switch self[idx] {
+            case 0x2D:  // -
+                self[idx] = 0x2B  // +
+            case 0x5F:  // _
+                self[idx] = 0x2F  // /
             default: break
             }
         }
@@ -33,11 +51,13 @@ private extension Data {
     /// Converts base64 encoded data to a base64-url encoded data.
     ///
     /// https://tools.ietf.org/html/rfc4648#page-7
-    mutating func base64URLEscape() {
-        for (i, byte) in enumerated() {
-            switch byte {
-            case 0x2B: self[self.index(self.startIndex, offsetBy: i)] = 0x2D
-            case 0x2F: self[self.index(self.startIndex, offsetBy: i)] = 0x5F
+    fileprivate mutating func base64URLEscape() {
+        for idx in self.indices {
+            switch self[idx] {
+            case 0x2B:  // +
+                self[idx] = 0x2D  // -
+            case 0x2F:  // /
+                self[idx] = 0x5F  // _
             default: break
             }
         }
@@ -47,7 +67,7 @@ private extension Data {
     /// Converts base64-url encoded data to a base64 encoded data.
     ///
     /// https://tools.ietf.org/html/rfc4648#page-7
-    func base64URLUnescaped() -> Data {
+    fileprivate func base64URLUnescaped() -> Data {
         var data = self
         data.base64URLUnescape()
         return data
@@ -56,7 +76,7 @@ private extension Data {
     /// Converts base64 encoded data to a base64-url encoded data.
     ///
     /// https://tools.ietf.org/html/rfc4648#page-7
-    func base64URLEscaped() -> Data {
+    fileprivate func base64URLEscaped() -> Data {
         var data = self
         data.base64URLEscape()
         return data
